requirements.txt
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.8.0
Flask-Talisman==1.1.0
pyserial==3.5
requests==2.32.3
python-dotenv==1.0.1
bcrypt==4.2.0
pytest==7.4.2